Installing java on EC2 in user data 3

+++ scripts/deploy_aws_ec2.sh
@@ -1,7 +1,22 @@
 #!/bin/bash
 
-#APP_JAR_FILE_NAME='translation-service-1.2.jar'
-APP_JAR_FILE_NAME=$1
+SERVICE_NAME=$1
+APP_JAR_FILE_NAME=$2
+
+
+if [ -z "$SERVICE_NAME" ]
+then
+      echo "Service name required (see EC2 tag Name value)"
+      exit 1;
+fi
+if [ -z "$APP_JAR_FILE_NAME" ]
+then
+      echo "Service JAR file name required"
+      exit 1;
+fi
+
+sudo mkdir -p "/var/log/bordozer/$SERVICE_NAME/"
+sudo chmod 777 . -R "/var/log/bordozer/$SERVICE_NAME/"
 
 echo "-- Deploy jar"
 kill -9 "$(lsof -t -i:8977)"
